safari-600.7-branch/LayoutTests/fast/regions/webkit-region-syntax-space.html

<!DOCTYPE html>
<html>
<head>
    <title>Spaces in @-webkit-region prelude - http://wkb.ug/123082</title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <style>@-webkit-region/**/#flow { #id {} }</style>
    <style>@-webkit-region/**/ #flow { #id {} }</style>
    <style>@-webkit-region /**/#flow { #id {} }</style>
    <style>@-webkit-region /**/ #flow { #id {} }</style>
</head>
<body>
    <div id="log"></div>
    <script>
        test(function(){
            assert_equals(document.styleSheets[0].cssRules.length, 1, "Rule valid and in CSSOM.");
        }, "Comment between @-symbol and flow name");

        test(function(){
            assert_equals(document.styleSheets[1].cssRules.length, 1, "Rule valid and in CSSOM.");
        }, "Comment + space between @-symbol and flow name");

        test(function(){
            assert_equals(document.styleSheets[2].cssRules.length, 1, "Rule valid and in CSSOM.");
        }, "Space + comment between @-symbol and flow name");

        test(function(){
            assert_equals(document.styleSheets[3].cssRules.length, 1, "Rule valid and in CSSOM.");
        }, "Space + comment + space between @-symbol and flow name");
    </script>
</body>
</html>
